backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
pandas
sentence-transformers
pillow
tqdm
mutagen
imagehash
ollama
timm
transformers
accelerate